@@ -1,5 +1,5 @@
 export const NavMenu = () => (
-  <React.Fragment>
+  <>
     <nav>
       <a href="#about">ABOUT</a>
       <a href="#team">TEAM</a>
@@ -36,5 +36,5 @@ export const NavMenu = () => (
         color: #000;
       }
     `}</style>
-  </React.Fragment>
+  </>
 )
